requirements.txt
openai
python-dotenv
flask
requests
sounddevice
numpy
faiss-cpu
faster-whisper
torch
scipy
webrtcvad